Refactor FinancialDashboard: replace mock data with dynamic balance fetching logic, update RequestsOverview to use RequestsOverviewContainer, and remove unused AlertButton component for cleaner code structure.
