+++ src/app/dashboard/balances/page.tsx
@@ -1,50 +1,9 @@
-import BalanceCard from "./components/BalanceCard/BalanceCard";
-import RequestsOverview from "./components/RequestsOverview/RequestsOverview";
-import BalanceCardSkeleton from "./components/BalanceCard/BalanceCardSkeleton";
+import RequestsOverviewContainer from "./components/RequestsOverview/RequestsOverviewContainer";
 import BalanceCardContainer from "./components/BalanceCard/BalanceCardContainer";
 
+export const dynamic = "force-dynamic";
+
 export default function FinancialDashboard() {
-  const mockPaidRequests = [
-    {
-      id: "1",
-      title: "Desarrollo Web Corporativo",
-      amount: 2500.0,
-      date: "2024-01-15",
-      client: "Empresa ABC",
-    },
-    {
-      id: "2",
-      title: "Consultoría IT",
-      amount: 1200.0,
-      date: "2024-01-10",
-      client: "StartupXYZ",
-    },
-    {
-      id: "3",
-      title: "Mantenimiento Sistema",
-      amount: 800.0,
-      date: "2024-01-08",
-      client: "TechCorp",
-    },
-  ];
-
-  const mockPendingRequests = [
-    {
-      id: "4",
-      title: "Diseño UX/UI",
-      amount: 1800.0,
-      date: "2024-01-20",
-      client: "Creative Agency",
-    },
-    {
-      id: "5",
-      title: "Integración API",
-      amount: 950.0,
-      date: "2024-01-18",
-      client: "FinTech Solutions",
-    },
-  ];
-
   return (
     <div className="min-h-screen w-full bg-[rgb(21,21,33)] p-4 md:p-6 lg:p-8">
       <div className="max-w-7xl mx-auto">
@@ -60,11 +19,7 @@ export default function FinancialDashboard() {
         <div className="grid grid-cols-1 lg:grid-cols-3 gap-6">
           <BalanceCardContainer />
           <div className="lg:col-span-2">
-            <RequestsOverview
-              paidRequests={mockPaidRequests}
-              pendingRequests={mockPendingRequests}
-              currency="€"
-            />
+            <RequestsOverviewContainer />
           </div>
         </div>
       </div>
